+++ src/app.tsx
@@ -15,6 +15,8 @@ import HomePage from './pages/home'
 import AuthPage from './pages/auth'
 import UnlockPage from './pages/unlock'
 import AddItemPage from './pages/item/add'
+import GroupPage from './pages/group'
+import GroupEditPage from './pages/group/edit'
 import ItemPage from './pages/item'
 import ItemEditPage from './pages/item/edit'
 import SettingsPage from './pages/settings'
@@ -64,6 +66,11 @@ function App () {
         return <HomePage />
       case 'favourites':
         return <HomePage />
+
+      case 'group':
+        return <GroupPage />
+      case 'group.edit':
+        return <GroupEditPage />
 
       case 'item':
         return <ItemPage />