+++ src/components/icons/folder.tsx
@@ -1,0 +1,21 @@
+import React from 'react'
+
+function FolderIcon () {
+  return (
+    <svg width="20px" height="16px" viewBox="0 0 20 16" version="1.1" xmlns="http://www.w3.org/2000/svg">
+      <g stroke="none" strokeWidth="1" fill="none" fillRule="evenodd">
+        <g transform="translate(-36.000000, -444.000000)" fill="currentColor" fillRule="nonzero">
+          <g transform="translate(20.000000, 409.000000)">
+            <g transform="translate(16.000000, 35.000000)">
+              <g transform="translate(0.000000, 0.000000)">
+                <path d="M8.71621053,2.94736842 C8.53726316,2.94736842 8.37557895,2.84084211 8.30484211,2.67621053 L7.71747368,1.30568421 C7.37810526,0.513684211 6.59915789,0 5.73726316,0 L2.15452632,0 C0.964631579,0 0,0.964631579 0,2.15452632 L0,2.89473684 L0,5.89473684 L0,13.3157895 C0,14.7983158 1.20168421,16 2.68421053,16 L16.6842105,16 C18.1667368,16 19.3684211,14.7983158 19.3684211,13.3157895 L19.3684211,5.63157895 C19.3684211,4.14905263 18.1667368,2.94736842 16.6842105,2.94736842 L8.71621053,2.94736842 Z" id="Path"></path>
+              </g>
+            </g>
+          </g>
+        </g>
+      </g>
+    </svg>
+  )
+}
+
+export default FolderIcon