+++ src/layouts/simple/index.tsx
@@ -3,7 +3,7 @@ import PropTypes from 'prop-types'
 
 import './_index.scss'
 
-const SimpleLayout: React.FC = ({ children }) => {
+const SimpleLayout: React.FC<{ children: React.ReactNode }> = ({ children }) => {
   return (
     <div className="layout -simple">
       <main>